requirements.txt
dotenv
requests
confluent-kafka
cloudant
ibmcloudant
ibm_watsonx_ai
langchain-ibm
load_dotenv
black
langgraph-swarm
pillow
langchain_community
langchain_core
chromadb
langchain_chroma
tiktoken
bs4
pypdf
twilio
streamlit